% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergeom.R
\name{hypergeom}
\alias{hypergeom}
\title{Computes the an approximation
of the probability mass function
of the hyper geometric distribution.}
\usage{
hypergeom(iterations, num_trials, total_y, total_obj_num)
}
\arguments{
\item{iterations}{the number of iterations to be performed to calculate averages}

\item{num_trials}{the number of trials (without substitution) to be performed}

\item{total_y}{the total successes}

\item{total_obj_num}{total number of objects in the set}
}
\description{
Try saying that 5 times fast.
}
\details{
Primarily performs the approximation
more accurately given higher 'iterations' values.
}
