% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{bootstrap sampling function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{iterations}

\item{x}{sample}

\item{fun}{function to be provided}

\item{alpha}{statistical alpha}

\item{cx}{typical cx}

\item{...}{remaining}
}
\description{
bootstrap sampling function
}
