% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Plots a dnorm function given mu and sigma, then
calculates the probability of P(Y<=A), and displays it}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{E(x) for dnorm}

\item{sigma}{sd for dnorm}

\item{a}{the x value to calculate the area until}
}
\description{
Plots a dnorm function given mu and sigma, then
calculates the probability of P(Y<=A), and displays it
}
